pyproject.toml
[project]
name = "mirage"
version = "0.1.0"
description = "A simple"
authors = [{ name = "limiteinductive", email = "benjamintrom@gmail.com" }]
dependencies = [
    "numpy>=1.24.4",
    "pillow>=10.0.1",
    "torch>=2.1.0",
]
readme = "README.md"
requires-python = ">= 3.11"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.9.1",
    "ruff>=0.0.292",
    "isort>=5.12.0",
    "docformatter>=1.7.5",
]

[tool.rye.scripts]
lint = "isort . && docformatter -r --in-place . && black . && ruff . --fix"

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 88
preview = true

[tool.docformatter]
black = true

[tool.isort]
profile = "black"
force_sort_within_sections = true
order_by_type = true 
combine_as_imports = true
